Move actionableItemCount to graphql redux
